+++ nessi/signal/tapering.py
@@ -89,7 +89,7 @@ def _cosine(n, ntap1, ntap2):
 
     return ftap
 
-def taper1d(dobs, ntap1, ntap2, min, type, axis=0):
+def taper1d(dobs, ntap1, ntap2, min=1.0, type='linear', axis=0):
     """
     Taper data.
     """
@@ -98,7 +98,7 @@ def taper1d(dobs, ntap1, ntap2, min, type, axis=0):
 
     # Get the number of dimensions of dobs
     if dobs.ndim == 1:
-        n = 1
+        n = len(dobs)
     if dobs.ndim == 2:
         if axis == 0:
             n = np.size(dobs, axis=0)
@@ -114,16 +114,14 @@ def taper1d(dobs, ntap1, ntap2, min, type, axis=0):
         ftap = _cosine(n, ntap1, ntap2)
 
     # Apply the taper function
-    if n == 1:
+    if dobs.ndim == 1:
         dobstaper[:] = dobs[:]*ftap[:]
     else:
         if axis == 0:
             for i in range(0, np.size(dobs, axis=1)):
                 dobstaper[:, i] = dobs[:, i]*ftap[:]
-            #print(dobs[:, 0],'\n',ftap[:],'\n',dobs[:, 0]*ftap[:])
         if axis == 1:
             for i in range(0, np.size(dobs, axis=0)):
                 dobstaper[i, :] = dobs[i, :]*ftap[:]
-            #print(dobs[0, :],'\n',ftap[:],'\n',dobs[0, :]*ftap[:])
 
     return dobstaper